+++ packages/ai/core/registry/index.ts
@@ -1,3 +1,7 @@
-export { experimental_customProvider } from './custom-provider';
-export * from './no-such-provider-error';
-export * from './provider-registry';
+export { customProvider, experimental_customProvider } from './custom-provider';
+export { NoSuchProviderError } from './no-such-provider-error';
+export {
+  createProviderRegistry,
+  experimental_createProviderRegistry,
+} from './provider-registry';
+export type { ProviderRegistryProvider } from './provider-registry';