+++ packages/ui-utils/src/data-stream-parts.ts
@@ -1,8 +1,8 @@
-import { LanguageModelV1FinishReason } from '@ai-sdk/provider';
 import {
-  ToolCall as ToolCall,
-  ToolResult as ToolResult,
-} from '@ai-sdk/provider-utils';
+  LanguageModelV1FinishReason,
+  LanguageModelV1Source,
+} from '@ai-sdk/provider';
+import { ToolCall, ToolResult } from '@ai-sdk/provider-utils';
 import { JSONValue } from './types';
 
 export type DataStreamString =
@@ -192,6 +192,7 @@ const finishMessageStreamPart: DataStreamPart<
   'finish_message',
   {
     finishReason: LanguageModelV1FinishReason;
+    // TODO v5 remove usage from finish event (only on step-finish)
     usage?: {
       promptTokens: number;
       completionTokens: number;
@@ -357,6 +358,95 @@ const reasoningStreamPart: DataStreamPart<'g', 'reasoning', string> = {
   },
 };
 
+const sourcePart: DataStreamPart<'h', 'source', LanguageModelV1Source> = {
+  code: 'h',
+  name: 'source',
+  parse: (value: JSONValue) => {
+    if (value == null || typeof value !== 'object') {
+      throw new Error('"source" parts expect a Source object.');
+    }
+
+    return {
+      type: 'source',
+      value: value as LanguageModelV1Source,
+    };
+  },
+};
+
+const redactedReasoningStreamPart: DataStreamPart<
+  'i',
+  'redacted_reasoning',
+  { data: string }
+> = {
+  code: 'i',
+  name: 'redacted_reasoning',
+  parse: (value: JSONValue) => {
+    if (
+      value == null ||
+      typeof value !== 'object' ||
+      !('data' in value) ||
+      typeof value.data !== 'string'
+    ) {
+      throw new Error(
+        '"redacted_reasoning" parts expect an object with a "data" property.',
+      );
+    }
+    return { type: 'redacted_reasoning', value: { data: value.data } };
+  },
+};
+
+const reasoningSignatureStreamPart: DataStreamPart<
+  'j',
+  'reasoning_signature',
+  { signature: string }
+> = {
+  code: 'j',
+  name: 'reasoning_signature',
+  parse: (value: JSONValue) => {
+    if (
+      value == null ||
+      typeof value !== 'object' ||
+      !('signature' in value) ||
+      typeof value.signature !== 'string'
+    ) {
+      throw new Error(
+        '"reasoning_signature" parts expect an object with a "signature" property.',
+      );
+    }
+    return {
+      type: 'reasoning_signature',
+      value: { signature: value.signature },
+    };
+  },
+};
+
+const fileStreamPart: DataStreamPart<
+  'k',
+  'file',
+  {
+    data: string; // base64 encoded data
+    mimeType: string;
+  }
+> = {
+  code: 'k',
+  name: 'file',
+  parse: (value: JSONValue) => {
+    if (
+      value == null ||
+      typeof value !== 'object' ||
+      !('data' in value) ||
+      typeof value.data !== 'string' ||
+      !('mimeType' in value) ||
+      typeof value.mimeType !== 'string'
+    ) {
+      throw new Error(
+        '"file" parts expect an object with a "data" and "mimeType" property.',
+      );
+    }
+    return { type: 'file', value: value as { data: string; mimeType: string } };
+  },
+};
+
 const dataStreamParts = [
   textStreamPart,
   dataStreamPart,
@@ -370,21 +460,19 @@ const dataStreamParts = [
   finishStepStreamPart,
   startStepStreamPart,
   reasoningStreamPart,
+  sourcePart,
+  redactedReasoningStreamPart,
+  reasoningSignatureStreamPart,
+  fileStreamPart,
 ] as const;
 
-type DataStreamParts =
-  | typeof textStreamPart
-  | typeof dataStreamPart
-  | typeof errorStreamPart
-  | typeof messageAnnotationsStreamPart
-  | typeof toolCallStreamPart
-  | typeof toolResultStreamPart
-  | typeof toolCallStreamingStartStreamPart
-  | typeof toolCallDeltaStreamPart
-  | typeof finishMessageStreamPart
-  | typeof finishStepStreamPart
-  | typeof startStepStreamPart
-  | typeof reasoningStreamPart;
+export const dataStreamPartsByCode = Object.fromEntries(
+  dataStreamParts.map(part => [part.code, part]),
+) as {
+  [K in (typeof dataStreamParts)[number]['code']]: (typeof dataStreamParts)[number];
+};
+
+type DataStreamParts = (typeof dataStreamParts)[number];
 
 /**
  * Maps the type of a stream part to its value type.
@@ -393,34 +481,7 @@ type DataStreamPartValueType = {
   [P in DataStreamParts as P['name']]: ReturnType<P['parse']>['value'];
 };
 
-export type DataStreamPartType =
-  | ReturnType<typeof textStreamPart.parse>
-  | ReturnType<typeof dataStreamPart.parse>
-  | ReturnType<typeof errorStreamPart.parse>
-  | ReturnType<typeof messageAnnotationsStreamPart.parse>
-  | ReturnType<typeof toolCallStreamPart.parse>
-  | ReturnType<typeof toolResultStreamPart.parse>
-  | ReturnType<typeof toolCallStreamingStartStreamPart.parse>
-  | ReturnType<typeof toolCallDeltaStreamPart.parse>
-  | ReturnType<typeof finishMessageStreamPart.parse>
-  | ReturnType<typeof finishStepStreamPart.parse>
-  | ReturnType<typeof startStepStreamPart.parse>
-  | ReturnType<typeof reasoningStreamPart.parse>;
-
-export const dataStreamPartsByCode = {
-  [textStreamPart.code]: textStreamPart,
-  [dataStreamPart.code]: dataStreamPart,
-  [errorStreamPart.code]: errorStreamPart,
-  [messageAnnotationsStreamPart.code]: messageAnnotationsStreamPart,
-  [toolCallStreamPart.code]: toolCallStreamPart,
-  [toolResultStreamPart.code]: toolResultStreamPart,
-  [toolCallStreamingStartStreamPart.code]: toolCallStreamingStartStreamPart,
-  [toolCallDeltaStreamPart.code]: toolCallDeltaStreamPart,
-  [finishMessageStreamPart.code]: finishMessageStreamPart,
-  [finishStepStreamPart.code]: finishStepStreamPart,
-  [startStepStreamPart.code]: startStepStreamPart,
-  [reasoningStreamPart.code]: reasoningStreamPart,
-} as const;
+export type DataStreamPartType = ReturnType<DataStreamParts['parse']>;
 
 /**
  * The map of prefixes for data in the stream
@@ -444,21 +505,11 @@ export const dataStreamPartsByCode = {
  * 6: {"tool_call": {"id": "tool_0", "type": "function", "function": {"name": "get_current_weather", "arguments": "{\\n\\"location\\": \\"Charlottesville, Virginia\\",\\n\\"format\\": \\"celsius\\"\\n}"}}}
  *```
  */
-export const DataStreamStringPrefixes = {
-  [textStreamPart.name]: textStreamPart.code,
-  [dataStreamPart.name]: dataStreamPart.code,
-  [errorStreamPart.name]: errorStreamPart.code,
-  [messageAnnotationsStreamPart.name]: messageAnnotationsStreamPart.code,
-  [toolCallStreamPart.name]: toolCallStreamPart.code,
-  [toolResultStreamPart.name]: toolResultStreamPart.code,
-  [toolCallStreamingStartStreamPart.name]:
-    toolCallStreamingStartStreamPart.code,
-  [toolCallDeltaStreamPart.name]: toolCallDeltaStreamPart.code,
-  [finishMessageStreamPart.name]: finishMessageStreamPart.code,
-  [finishStepStreamPart.name]: finishStepStreamPart.code,
-  [startStepStreamPart.name]: startStepStreamPart.code,
-  [reasoningStreamPart.name]: reasoningStreamPart.code,
-} as const;
+export const DataStreamStringPrefixes = Object.fromEntries(
+  dataStreamParts.map(part => [part.name, part.code]),
+) as {
+  [K in DataStreamParts['name']]: (typeof dataStreamParts)[number]['code'];
+};
 
 export const validCodes = dataStreamParts.map(part => part.code);